exportPa.py fully functional. TODOs pending...
Pending to add CLI arguments processing. At least two arguments: -s
and --skip-main-sf indicating that the first subflow is not the
main. Very useful for error recovering.

Pending fix the waitDisappearAll(). Sometimes it's not recognizing
when an image disappears. Converting it to a parallel version with
"concurrent.futures" might be a good option (analogue to the
waitAny()).